Adds selected comment folding functionality
Selected comment folding, "ctrl+alt+c", can now be used to just fold
comments that are currently selected (either the cursor is within the
comment or a portion of the comment is selected).

The intialization of self.comments was changed so append could be used
in the new method for finding selected comments.  I don't think that
anything else relied upon this, but I thought it was worth noting.
Beyond that some of the functions were taken out of the init routine
just because they weren't common anymore.